Implemented the attack system and added all monster targets
Made an extensible attack system that allows every attack to have extra
features (each attack is a ScriptableObject that is turned into an asset
by a custom editor script), populated every monster's new attack list
with it's respective attacks, and created different targets for every
monster, so every monster is now up and running (already in the Game
scene). Only thing need now is to make it so that the HUD sends the
target object to the attack method and its perfect.